ml/requirements.txt
flask
pymongo
scikit-learn
pandas
numpy
python-dotenv
dnspython
